+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Release 5.6.5
  * Credentials: Safe decoding of images in credentials
+ * Credentials: Update EU PID to allow for LocalDate or Instant in expiry, issuance date
 
 # Release 5.6.4
  * Add: Koin Dependency Injection